Add catURL for intellegently concatenating urls
Prior the script would add the baseurl onto the relative path to
somewhere intern on the website no matter what. The problem occured when
the website would link to somewhere externally, where the script would
add the domain of the first website onto the url for the other, and
thereby ruining the result. This has been fixed by checking whether the
path scraped already contains a domain and specifications for the http
standard, and only adding the rooturl if it doesn't.